Branched from $/RapidDevBookCode/LLBL Pro v4.1

+++ AW.Winforms.Helpers/MostRecentlyUsedHandler/MRUHandler.cs
@@ -24,71 +24,55 @@ using System.Windows.Forms;
 
 namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
 {
-  ///
   /// Enumerated type for the MRUStyle property
-  ///
-  ///
+  /// 
+  /// 
   /// Popup
   /// The Most Recently Used list is shown as a pop up menu.
-  ///
-  ///
+  /// 
+  /// 
   /// Inline
   /// The Most Recenly Used list is included in the menu where the desginated item would go.
-  ///
-  ///
-  ///
   public enum MRUStyle
   {
     Popup,
     Inline
-  } ;
+  };
 
-  ///
   /// Enumerated type for the defining the persistent storage location of the Most Recently Used history
-  ///
-  ///
+  /// 
+  /// 
   /// File
   /// The Most Recently Used list persisted to the local file system.
-  ///
-  ///
+  /// 
+  /// 
   /// Registry
   /// The Most Recenly Used list is stored in the win32 registry.
-  ///
-  ///
+  /// 
+  /// 
   /// IsolatedStorage
   /// The Most Recenly Used list persisted in the .Net Isolated Storage.
-  ///
-  ///
-  ///
   public enum StorageStyle
   {
     File,
     Registry,
     IsolatedStorage
-  } ;
+  };
 
-  ///
   /// The delegate definition for the Most Recently Used menu click event
-  ///
   public delegate void MRUItemClickedHandler(object sender, MRUItemClickedEventArgs e);
 
-  ///
   /// Summary description for MRUHandler.
-  ///
   [ToolboxBitmap(typeof (MRUHandler), "MRUHandler.bmp")]
-  public class MRUHandler: Component, ISupportInitialize
+  public class MRUHandler : Component, ISupportInitialize
   {
     #region private static definitions
 
-    ///
     /// Defines the Most Recently Used file definition
-    ///
-    private static readonly string MRU_EXTENSION = ".mru";
+    private const string MRU_EXTENSION = ".mru";
 
-    ///
     /// Defines the registry key base name that is used to generate the key for each mru entry
-    ///
-    private static readonly string REGISTRY_VALUE_NAME_BASE = "mru_entry_";
+    private const string REGISTRY_VALUE_NAME_BASE = "mru_entry_";
 
     #endregion
 
@@ -101,7 +85,9 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       container.Add(this);
     }
 
-    public MRUHandler() {}
+    public MRUHandler()
+    {
+    }
 
     #endregion
 
@@ -130,26 +116,24 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
 
     #endregion
 
-    ///
     /// Is the internal menu click handler for all of the Most Recently Used menu items. Based on the menuitem
     /// the file name is retrieved from the Hashtable. The display text of the item will not be a usable file name.
     /// For instance it has a number in front and the text could have been compressed by the 'ShowShortItem' property
-    ///
+    /// 
     /// Object that raised the event
     /// Event information
-    protected void onMenuItemClick(Object sender, EventArgs e)
+    private void onMenuItemClick(Object sender, EventArgs e)
     {
-      var item = (ToolStripMenuItem)sender;
+      var item = (ToolStripMenuItem) sender;
       if (MRUItemClicked != null)
       {
-        var fileName = (string)_itemMap[item];
+        var fileName = (string) _itemMap[item];
         MRUItemClicked(this, new MRUItemClickedEventArgs(fileName));
       }
     }
 
-    ///
     /// Add a file path name to the Most Recently Used history
-    ///
+    /// 
     /// path to be added
     public void AddRecentlyUsedFile(string fileName)
     {
@@ -163,10 +147,8 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
 
     #region properties
 
-    ///
     /// Property to determine the menuitem that becomes the place holder for the
     /// Most Recently Used Menu items.
-    ///
     [Description("Specifies the menu item that is used as the MRU list parent.")]
     public ToolStripMenuItem MruItem
     {
@@ -179,9 +161,7 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       }
     }
 
-    ///
     /// Property that specifies what name is used in the storage of the MRU history
-    ///
     [Description("Specifies the name to be used to generate the file name for persistence.")]
     public string StorageName
     {
@@ -194,9 +174,7 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       }
     }
 
-    ///
     /// Property that determins the style of the MRU list at runtime
-    ///
     [Description("Specifies the style of the MRU list."),
      DefaultValue(MRUStyle.Popup)]
     public MRUStyle MRUStyle
@@ -210,9 +188,7 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       }
     }
 
-    ///
     /// Property to determine if the MRU menu item will show the truncated text.
-    ///
     [Description("Specifies whether the file name in the MRU entry should be truncated."),
      DefaultValue(false)]
     public bool ShowShortItem
@@ -226,12 +202,10 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       }
     }
 
-    ///
     /// Property to determine if separators will be automatically placed arround inline MRU entries when necessery.
-    ///
     [Description("Specifies whether separators will be automatically placed arround inline MRU entries when necessery."),
      DefaultValue(false)] //default value is "False" to maintain backwards compatibility
-      public bool AutoSeparators //Birb: added
+    public bool AutoSeparators //Birb: added
     {
       get { return _autoSeparators; }
       set
@@ -242,9 +216,7 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       }
     }
 
-    ///
     /// Property to determine the type of persistent storage that should be used to store the MRU history.
-    ///
     [Description("Specifies the type of persistent storage to use to store the MRU history."),
      DefaultValue(StorageStyle.IsolatedStorage)]
     public StorageStyle StorageStyle
@@ -258,9 +230,7 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       }
     }
 
-    ///
     /// Property to determine the length of the MRU history.
-    ///
     [Description("Specifies the length of the MRU history file."),
      DefaultValue(10)]
     public int MaxFiles
@@ -276,9 +246,7 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       }
     }
 
-    ///
     /// Property to determine the length of text to be shown if the ShowShortItem property is set to true.
-    ///
     [Description("Specifies the length of the MRU menu text when the ShowShortItem property is true."),
      DefaultValue(40)]
     public int DisplayLength
@@ -296,9 +264,7 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
 
     #region Component overrides
 
-    ///
     /// Clean up any resources being used.
-    ///
     protected override void Dispose(bool disposing)
     {
       if (_mruIsDirty)
@@ -334,21 +300,18 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
 
       // Read the stream from Isolated Storage.
       Stream stream = new IsolatedStorageFileStream(MRUFile(), FileMode.OpenOrCreate, isoStore);
-      if (stream != null)
-        try
-        {
-          // DeSerialize the Hashtable from stream.
-          IFormatter formatter = new BinaryFormatter();
-          var appData = (ArrayList)formatter.Deserialize(stream);
-          return appData;
-        }
-        finally
-        {
-          // We are done with it.
-          stream.Close();
-        }
-      else
-        return new ArrayList();
+      try
+      {
+        // DeSerialize the Hashtable from stream.
+        IFormatter formatter = new BinaryFormatter();
+        var appData = (ArrayList) formatter.Deserialize(stream);
+        return appData;
+      }
+      finally
+      {
+        // We are done with it.
+        stream.Close();
+      }
     }
 
     private void saveToStorage()
@@ -357,17 +320,16 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       var isoStore = IsolatedStorageFile.GetStore(IsolatedStorageScope.User | IsolatedStorageScope.Assembly, null, null);
       Stream stream = new IsolatedStorageFileStream(MRUFile(), FileMode.Create, isoStore);
 
-      if (stream != null)
-        try
-        {
-          // Serialize the Hashtable into the IsolatedStorage.
-          IFormatter formatter = new BinaryFormatter();
-          formatter.Serialize(stream, MRUList());
-        }
-        finally
-        {
-          stream.Close();
-        }
+      try
+      {
+        // Serialize the Hashtable into the IsolatedStorage.
+        IFormatter formatter = new BinaryFormatter();
+        formatter.Serialize(stream, MRUList());
+      }
+      finally
+      {
+        stream.Close();
+      }
     }
 
     #endregion
@@ -411,24 +373,21 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
     private ArrayList loadFromRegistry()
     {
       var key = Application.UserAppDataRegistry;
+      if (key == null) return new ArrayList();
       var mruKey = key.OpenSubKey(_fileName, true);
-      if (mruKey != null)
+      if (mruKey == null) return new ArrayList();
+      // does it exist, if so load the values
+      var mruCount = mruKey.GetValue("mru_count");
+      var mruSize = (int) mruCount;
+      var mru = new ArrayList(mruSize);
+      for (var i = 0; i < mruSize; i++)
       {
-        // does it exist, if so load the values
-        var mruCount = mruKey.GetValue("mru_count");
-        var mruSize = (int)mruCount;
-        var mru = new ArrayList(mruSize);
-        for (var i = 0; i < mruSize; i++)
-        {
-          var sb = new StringBuilder(REGISTRY_VALUE_NAME_BASE.Length + 10);
-          sb.Append(REGISTRY_VALUE_NAME_BASE).Append(i);
-          mru.Add(mruKey.GetValue(sb.ToString()));
-        }
-
-        return mru;
+        var sb = new StringBuilder(REGISTRY_VALUE_NAME_BASE.Length + 10);
+        sb.Append(REGISTRY_VALUE_NAME_BASE).Append(i);
+        mru.Add(mruKey.GetValue(sb.ToString()));
       }
-      else
-        return new ArrayList();
+
+      return mru;
     }
 
     private void saveToRegistry()
@@ -438,12 +397,14 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       var key = Application.UserAppDataRegistry;
 
       // no child keys anyway but also say don't throw
+      if (key == null) return;
       key.DeleteSubKey(_fileName, false); // so we can replace all the duplicates
 
       var mruKey = key.CreateSubKey(_fileName);
       /*-------------*/
 
       var counter = 0;
+      if (mruKey == null) return;
       mruKey.SetValue("mru_count", MRUList().Count);
       foreach (String str in MRUList())
       {
@@ -457,9 +418,7 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
 
     #region menu item builder functions
 
-    ///
     /// Rebuilds the Most Rcently Used menu base on the value of the MRUStyle property
-    ///
     private void RebuildMenu()
     {
       if (_style == MRUStyle.Popup)
@@ -468,9 +427,7 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
         RebuildInlineMenu();
     }
 
-    ///
     /// Rebuild the popup version of the most recently used list
-    ///
     private void RebuildPopupMenu() //Birb: added for better code readability (since original code already had a separate "RebuildInlineMenu")
     {
       _mruItem.DropDownItems.Clear(); //Birb: renamed MenuItems to DropDownItems
@@ -485,9 +442,7 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       }
     }
 
-    ///
     /// Rebuild the inline version of the most recently used list
-    ///
     private void RebuildInlineMenu()
     {
       foreach (ToolStripMenuItem item in _itemMap.Keys)
@@ -502,7 +457,7 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       _mruItem.Enabled = false;
 
       var offset = 1;
-      var parentMenu = (ToolStripMenuItem)_mruItem.OwnerItem; // get our parent //Birb: renamed Parent to OwnerItem
+      var parentMenu = (ToolStripMenuItem) _mruItem.OwnerItem; // get our parent //Birb: renamed Parent to OwnerItem
       var parentItems = parentMenu.DropDownItems; //Birb: used instead of parentMenu.MenuItems
       var MRUMenuIndex = /*_mruItem.Index*/ parentItems.IndexOf(_mruItem); // where are we in the chain //Birb: changed the code to find the _mruItem index inside its parent since there's no Index property at ToolStripItem
 
@@ -520,7 +475,7 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
         //Birb: "append" separator to mru item if not last item and not existing already
         if ((MRUMenuIndex != parentItems.Count - 1) && !(parentItems[MRUMenuIndex + 1] is ToolStripSeparator))
           parentMenu.DropDownItems.Insert(MRUMenuIndex + 1, new ToolStripSeparator());
-          //don't update the mru item index, we appended something (didn't prepend anything)
+        //don't update the mru item index, we appended something (didn't prepend anything)
       }
 
       foreach (string str in MRUList())
@@ -532,12 +487,10 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       }
     }
 
-    ///
     /// Builds a menu item for the path entry
-    ///
+    /// 
     /// Position in list
     /// path
-    ///
     private ToolStripMenuItem buildMenuItem(int position, string text)
     {
       var sb = new StringBuilder(text.Length + 10);
@@ -560,14 +513,13 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       return item;
     }
 
-    ///
     /// Truncate a path to fit within a certain number of characters
     /// by replacing path components with ellipses.
-    ///
+    /// 
     /// This solution is provided by CodeProject and GotDotNet C# expert
     /// Richard Deeming.
-    ///
-    ///
+    /// 
+    /// 
     /// Long file name
     /// Maximum length
     /// Truncated file name
@@ -576,18 +528,16 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       var pszOut = new StringBuilder(maxLen + maxLen + 2); // for safety
       if (PathCompactPathEx(pszOut, longName, maxLen, 0))
         return pszOut.ToString();
-      else
-        return longName;
+      return longName;
     }
 
     #endregion
 
     #region internal helper functions
 
-    ///
     /// Generates a standard file name to be used to persist the mru history.
     /// The formation is product name_file name property.mru
-    ///
+    /// 
     /// A string representing the file name to be used to store the mru history/returns>
     private string MRUFile()
     {
@@ -596,9 +546,8 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
       return sb.ToString();
     }
 
-    ///
     /// Loads the mru history from the defined storage location.
-    ///
+    /// 
     /// An array containing the MRU history
     private ArrayList MRUList()
     {
@@ -628,21 +577,17 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
 
     #region ISupportInitialize Implementation
 
-    ///
     /// Implementation of the BeginInit Method. Set a flag so that we are initializing the
     /// component by the framework. This will bypass all the run time error check on setting
     /// the parameters.
-    ///
     public void BeginInit()
     {
       _isInitializing = true;
     }
 
-    ///
     /// Implementation of the EndInit Method. When we this method is called, the framework
     /// has completed initializing the component with the values that were set in the designer.
     /// We now set the flag so that run time parameter setting is basically disabled.
-    ///
     public void EndInit()
     {
       _isInitializing = false;
@@ -657,10 +602,8 @@ namespace AW.Winforms.Helpers.MostRecentlyUsedHandler
     #endregion
   }
 
-  ///
   /// The event arguments for the generated Most Recently Used menu click event
-  ///
-  public class MRUItemClickedEventArgs: EventArgs
+  public class MRUItemClickedEventArgs : EventArgs
   {
     private readonly string _file;
